quore_clone/quore_clone/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
